myStockroom/requirements.txt
ipython
asgiref
attrs
diff-match-patch
Django >=4.1.6,<=5
djangorestframework
markdown
django-filter
gunicorn
celery
django-celery-beat
redis
hiredis
flower
django-adaptors
django-appconf
django-cors-headers
django-crispy-forms
django-bootstrap5
crispy-bootstrap5
Pillow
psycopg2
psycopg2-binary
python-dateutil
pytz
six
sqlparse
django-debug-toolbar
pytest
pytest-django
pytest-xdist
pytest-cov